// app/(employees)/shifts/types.ts
export interface ShiftDetailsProps {
  id: string,
  shiftType: string;
  totalEarnings: string;
  date: string;
  time: string;
  endTime: string;
  hourlyRate: string;
  numberOfHours: string;
  numberOfOpenings: string;
  location: string;
  description: string;
  employee: string;
  employer: string;
  employerId: string;
  employeeId: string;
  backgroundImage: string;
}